+++ yb-voyager/cmd/live_migration_cdc_partition_strategy.go
@@ -1,0 +1,303 @@
+/*
+Copyright (c) YugabyteDB, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package cmd
+
+import (
+	"fmt"
+	"strings"
+
+	goerrors "github.com/go-errors/errors"
+	log "github.com/sirupsen/logrus"
+
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/metadb"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/namereg"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/tgtdb"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils/sqlname"
+)
+
+/*
+prepareCdcPartitionKey validates overrides against the import table list, resolves
+per-table strategies (global --cdc-partition-key + --cdc-partition-key-overrides +
+auto/expr-UK rules), and persists TableToCDCPartitioningStrategyMap.
+
+It is intentionally called before snapshot import so bad configs fail fast.
+On resume (map already in metaDB) this is a no-op.
+*/
+func prepareCdcPartitionKey(tableNames []sqlname.NameTuple) error {
+	if importerRole != TARGET_DB_IMPORTER_ROLE || !changeStreamingIsEnabled(importType) || sourceDBType != POSTGRESQL {
+		return nil
+	}
+
+	importDataStatus, err := metaDB.GetImportDataStatusRecord()
+	if err != nil {
+		return fmt.Errorf("error getting import data status record for cdc-partition-key: %w", err)
+	}
+	if importDataStatus == nil {
+		return goerrors.Errorf("import data status record not found")
+	}
+	if importDataStatus.TableToCDCPartitioningStrategyMap != nil {
+		//TODO: see if we should validation here after computing map again witht he keys and overrides and then check for any mismatch
+		log.Infof("cdc partition key already prepared in metadb; skipping recompute")
+		return nil
+	}
+
+	_, err = computeAndPersistCdcPartitioningStrategyPerTable(tableNames)
+	return err
+}
+
+/*
+getCdcPartitioningStrategyPerTable loads the per-table CDC partition strategy for streaming.
+
+For target PG→YB live import the map is prepared before snapshot via prepareCdcPartitionKey.
+Non-target and Oracle source paths force PARTITION_BY_TABLE in-memory (not persisted).
+
+TODO: handle upgrade scenario for PG/Oracle pk->table change
+*/
+func getCdcPartitioningStrategyPerTable(tableNames []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, string], error) {
+	tableToPartitioningStrategyMap := utils.NewStructMap[sqlname.NameTuple, string]()
+
+	if importerRole != TARGET_DB_IMPORTER_ROLE {
+		//For PG/ORacle source/source-replica, using partitioning by table since there won't be any huge difference in
+		// performance between the two strategies for single node databases like PG/Oracle
+		//and Parititon by table is better from data correctness perspective
+		for _, t := range tableNames {
+			tableToPartitioningStrategyMap.Put(t, PARTITION_BY_TABLE)
+		}
+		return tableToPartitioningStrategyMap, nil
+	}
+
+	if sourceDBType != POSTGRESQL {
+		//Oracle sources do not support unique-key conflict detection during live migration
+		//(we do not fetch unique indexes for Oracle). Force PARTITION_BY_TABLE so that all
+		//events of a table run sequentially on a single channel, which makes unique-key
+		//conflicts impossible and hence conflict detection unnecessary.
+		//anything other than PG is not supported for conflict and hence we force table partitioning
+		for _, t := range tableNames {
+			tableToPartitioningStrategyMap.Put(t, PARTITION_BY_TABLE)
+		}
+		return tableToPartitioningStrategyMap, nil
+	}
+
+	importDataStatus, err := metaDB.GetImportDataStatusRecord()
+	if err != nil {
+		return nil, fmt.Errorf("error getting cdc partitioning strategy: %w", err)
+	}
+	if importDataStatus == nil {
+		return nil, goerrors.Errorf("import data status record not found")
+	}
+	if importDataStatus.TableToCDCPartitioningStrategyMap != nil {
+		log.Infof("cdc partitioning strategy found in metadb: %v, strategy: %v", metadb.IMPORT_DATA_STATUS_KEY, importDataStatus.TableToCDCPartitioningStrategyMap)
+		for tableName, strategy := range importDataStatus.TableToCDCPartitioningStrategyMap {
+			tuple, err := namereg.NameReg.LookupTableName(tableName)
+			if err != nil {
+				return nil, fmt.Errorf("error looking up table name: %w", err)
+			}
+			tableToPartitioningStrategyMap.Put(tuple, strategy)
+		}
+		for _, t := range tableNames {
+			if _, ok := tableToPartitioningStrategyMap.Get(t); !ok {
+				return nil, goerrors.Errorf("cdc partitioning strategy not found for table: %s", t.ForKey())
+			}
+		}
+		return tableToPartitioningStrategyMap, nil
+	}
+	return nil, goerrors.Errorf("cdc partitioning strategy not found in metadb")
+}
+
+// resolveEffectiveCdcPartitionKeys applies global strategy, then per-table overlays,
+// then rejects effective pk for expression-UK tables. Pure helper for unit tests.
+func resolveEffectiveCdcPartitionKeys(
+	tableNames []sqlname.NameTuple,
+	globalKey string,
+	overrides *utils.StructMap[sqlname.NameTuple, string],
+	exprUKSet *utils.StructMap[sqlname.NameTuple, bool],
+	isAMP bool,
+) (*utils.StructMap[sqlname.NameTuple, string], error) {
+	result := utils.NewStructMap[sqlname.NameTuple, string]()
+	if overrides == nil {
+		overrides = utils.NewStructMap[sqlname.NameTuple, string]()
+	}
+	if exprUKSet == nil {
+		exprUKSet = utils.NewStructMap[sqlname.NameTuple, bool]()
+	}
+
+	switch globalKey {
+	case "auto":
+		if isAMP {
+			// yb-amp is a single-node PostgreSQL-compatible compute (no YB
+			// tablets / colocation), so — exactly like the PG/Oracle
+			// source/source-replica case — PARTITION_BY_TABLE has no real
+			// throughput downside and is safer for correctness. It also sidesteps
+			// getExpressionUniqueIndexTables(), which is implemented only on the
+			// TargetYugabyteDB driver.
+			for _, t := range tableNames {
+				result.Put(t, PARTITION_BY_TABLE)
+			}
+		} else {
+			for _, t := range tableNames {
+				if _, ok := exprUKSet.Get(t); ok {
+					result.Put(t, PARTITION_BY_TABLE)
+				} else {
+					result.Put(t, PARTITION_BY_PK)
+				}
+			}
+		}
+	default:
+		for _, t := range tableNames {
+			result.Put(t, globalKey)
+		}
+	}
+
+	err := overrides.IterKV(func(t sqlname.NameTuple, strategy string) (bool, error) {
+		result.Put(t, strategy)
+		return true, nil
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error applying cdc-partition-key-overrides: %w", err)
+	}
+
+	for _, t := range tableNames {
+		strategy, _ := result.Get(t)
+		if strategy == PARTITION_BY_PK {
+			if _, isExprUK := exprUKSet.Get(t); isExprUK {
+				return nil, goerrors.Errorf("cdc-partition-key pk is not allowed for table %s because it has an expression-based unique index; use table (via --cdc-partition-key or --cdc-partition-key-overrides)", t.ForOutput())
+			}
+		}
+	}
+	return result, nil
+}
+
+// resolveCdcPartitionKeyOverrides looks up override table names in namereg and
+// validates each is present in importTableList. Returns a map keyed by NameTuple.
+func resolveCdcPartitionKeyOverrides(rawOverrides map[string]string, importTableList []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, string], error) {
+	resolved := utils.NewStructMap[sqlname.NameTuple, string]()
+	if len(rawOverrides) == 0 {
+		return resolved, nil
+	}
+
+	importTableSet := utils.NewStructMap[sqlname.NameTuple, bool]()
+	for _, t := range importTableList {
+		importTableSet.Put(t, true)
+	}
+
+	for tableSpec, strategy := range rawOverrides {
+		tuple, err := namereg.NameReg.LookupTableName(tableSpec)
+		if err != nil {
+			return nil, goerrors.Errorf("cdc-partition-key-overrides: table %q not found in name registry: %w", tableSpec, err)
+		}
+		if _, ok := importTableSet.Get(tuple); !ok {
+			return nil, goerrors.Errorf("cdc-partition-key-overrides: table %q is not in the import table list", tableSpec)
+		}
+		// Detect duplicates on the resolved NameTuple so different spellings of the
+		// same table (casing/quoting/schema-qualification) don't silently overwrite.
+		if existing, ok := resolved.Get(tuple); ok && !strings.EqualFold(existing, strategy) {
+			return nil, goerrors.Errorf("cdc-partition-key-overrides: table %q (resolved to %s) specified multiple times with conflicting strategies %q and %q",
+				tableSpec, tuple.ForOutput(), existing, strategy)
+		}
+		resolved.Put(tuple, strategy)
+	}
+	return resolved, nil
+}
+
+func checkIfNeedsExprUKCheck(cdcPartitionKey string, overrides *utils.StructMap[sqlname.NameTuple, string]) (bool, error) {
+	// Collect expression-UK tables whenever we need them for auto resolution or pk-on-expr-UK validation.
+	needsExprUKCheck := cdcPartitionKey == "auto" || cdcPartitionKey == PARTITION_BY_PK
+	if !needsExprUKCheck {
+		err := overrides.IterKV(func(_ sqlname.NameTuple, strategy string) (bool, error) {
+			if strategy == PARTITION_BY_PK {
+				needsExprUKCheck = true
+				return false, nil
+			}
+			return true, nil
+		})
+		if err != nil {
+			return false, err
+		}
+	}
+	return needsExprUKCheck, nil
+}
+
+// computeAndPersistCdcPartitioningStrategyPerTable resolves global + overrides + auto/expr-UK
+// rules and writes TableToCDCPartitioningStrategyMap to metaDB.
+func computeAndPersistCdcPartitioningStrategyPerTable(tableNames []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, string], error) {
+	rawOverrides, err := parseCdcPartitionKeyOverrides(cdcPartitionKeyOverrides)
+	if err != nil {
+		return nil, err
+	}
+	overrides, err := resolveCdcPartitionKeyOverrides(rawOverrides, tableNames)
+	if err != nil {
+		return nil, err
+	}
+
+	needsExprUKCheck, err := checkIfNeedsExprUKCheck(cdcPartitionKey, overrides)
+	if err != nil {
+		return nil, err
+	}
+
+	var expressionUniqueIndexTables []sqlname.NameTuple
+	if needsExprUKCheck {
+		expressionUniqueIndexTables, err = getExpressionUniqueIndexTables(tableNames)
+		if err != nil {
+			return nil, fmt.Errorf("error getting expression unique index tables: %w", err)
+		}
+	}
+	exprUKSet := utils.NewStructMap[sqlname.NameTuple, bool]()
+	for _, t := range expressionUniqueIndexTables {
+		exprUKSet.Put(t, true)
+	}
+
+	tableToPartitioningStrategyMap, err := resolveEffectiveCdcPartitionKeys(
+		tableNames, cdcPartitionKey, overrides, exprUKSet, tconf.TargetDBType == YUGABYTEDB_AMP)
+	if err != nil {
+		return nil, err
+	}
+
+	metadbMap := make(map[string]string)
+	err = tableToPartitioningStrategyMap.IterKV(func(key sqlname.NameTuple, value string) (bool, error) {
+		metadbMap[key.ForKey()] = value
+		return true, nil
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error building cdc partitioning strategy map: %w", err)
+	}
+	err = metaDB.UpdateImportDataStatusRecord(func(obj *metadb.ImportDataStatusRecord) {
+		obj.TableToCDCPartitioningStrategyMap = metadbMap
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error updating cdc partitioning strategy in metadb: %w", err)
+	}
+	log.Infof("updated cdc partitioning strategy in metadb: %v with values: %v", metadb.IMPORT_DATA_STATUS_KEY, metadbMap)
+	return tableToPartitioningStrategyMap, nil
+}
+
+func getExpressionUniqueIndexTables(tableNames []sqlname.NameTuple) ([]sqlname.NameTuple, error) {
+	if tconf.TargetDBType != YUGABYTEDB {
+		return nil, nil
+	}
+	yb, ok := tdb.(*tgtdb.TargetYugabyteDB)
+	if !ok {
+		return nil, goerrors.Errorf("target db is not a YugabyteDB")
+	}
+
+	//returns a list of catalog table names, in case partitions it return catalog leaf partitions names and root table names
+	expressionUniqueIndexTables, err := yb.GetTablesHavingExpressionUniqueIndexes(tableNames, true)
+	if err != nil {
+		return nil, fmt.Errorf("error getting tables having expression or normal unique indexes: %w", err)
+	}
+
+	return expressionUniqueIndexTables, nil
+}
